+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Ability to define user groups in BasicAuth
 - Group-based permission and protection functions
 
+### Fixed
+- Fix when looking for callback inputs that are not in the right format when checking for whitelisted routes
+
 ## [2.2.0] - 2024-02-05
 ### Added
 - Possibility to whitelist routes with the `add_public_routes` utility function, the routes should follow Flask route syntax